Refactor: move Interval into utils
